refactor: #131 全件削除ボタンのコンポーネント化
